memes: lots of animals here.

+++ userbot/plugins/memes.py
@@ -257,7 +257,7 @@ PASTAMOJIS = [
                   outgoing=True,
                   regex="shibe$")
 async def shibes(event: NewMessage.Event) -> None:
-    """Get random pictures of Shibes."""
+    """Get random pictures of shibes."""
     shibe = await _request('http://shibe.online/api/shibes')
     if not shibe:
         await event.answer("`Couldn't fetch a shibe for you :(`")
@@ -273,7 +273,7 @@ async def shibes(event: NewMessage.Event) -> None:
 
 @client.onMessage(command=("cat", plugin_category),
                   outgoing=True,
-                  regex="cat$")
+                  regex=r"(cat|🐈)$")
 async def cats(event: NewMessage.Event) -> None:
     """Get random pictures of cats."""
     shibe = await _request('http://shibe.online/api/cats')
@@ -291,7 +291,7 @@ async def cats(event: NewMessage.Event) -> None:
 
 @client.onMessage(command=("bird", plugin_category),
                   outgoing=True,
-                  regex="bird$")
+                  regex=r"(bird|🐦)$")
 async def birds(event: NewMessage.Event) -> None:
     """Get random pictures of birds."""
     shibe = await _request('http://shibe.online/api/birds')
@@ -302,6 +302,87 @@ async def birds(event: NewMessage.Event) -> None:
     _, json = shibe
     try:
         await event.answer(file=json[0], reply_to=event.reply_to_msg_id)
+        await event.delete()
+    except rpcerrorlist.TimeoutError:
+        await event.answer("`Event timed out!`")
+
+
+@client.onMessage(command=("dog", plugin_category),
+                  outgoing=True,
+                  regex=r"(?:🐕|dog)(?: |$)(\w+)?(?: |$)(\w+)?")
+async def dogs(event: NewMessage.Event) -> None:
+    """Get random pictures of dogs."""
+    breed = event.matches[0].group(1)
+    subbreed = event.matches[0].group(2)
+    if breed and subbreed:
+        url = f"https://dog.ceo/api/breed/{breed}/{subbreed}/images/random"
+    elif breed:
+        url = f"https://dog.ceo/api/breed/{breed}/images/random"
+    else:
+        url = "https://dog.ceo/api/breeds/image/random"
+    dog = await _request(url)
+    if not dog:
+        await event.answer("`Couldn't fetch a dog for you :(`")
+        return
+
+    _, json = dog
+    try:
+        await event.answer(file=json['message'],
+                           reply_to=event.reply_to_msg_id)
+        await event.delete()
+    except rpcerrorlist.TimeoutError:
+        await event.answer("`Event timed out!`")
+
+
+@client.onMessage(command=("fox", plugin_category),
+                  outgoing=True,
+                  regex=r"(fox|🦊)$")
+async def foxes(event: NewMessage.Event) -> None:
+    """Get random pictures of foxes."""
+    fox = await _request('https://some-random-api.ml/img/fox')
+    if not fox:
+        await event.answer("`Couldn't fetch a fox for you :(`")
+        return
+
+    _, json = fox
+    try:
+        await event.answer(file=json['link'], reply_to=event.reply_to_msg_id)
+        await event.delete()
+    except rpcerrorlist.TimeoutError:
+        await event.answer("`Event timed out!`")
+
+
+@client.onMessage(command=("panda", plugin_category),
+                  outgoing=True,
+                  regex=r"(panda|🐼)$")
+async def pandas(event: NewMessage.Event) -> None:
+    """Get random pictures of pandas."""
+    panda = await _request('https://some-random-api.ml/img/panda')
+    if not panda:
+        await event.answer("`Couldn't fetch a panda for you :(`")
+        return
+
+    _, json = panda
+    try:
+        await event.answer(file=json['link'], reply_to=event.reply_to_msg_id)
+        await event.delete()
+    except rpcerrorlist.TimeoutError:
+        await event.answer("`Event timed out!`")
+
+
+@client.onMessage(command=("redpanda", plugin_category),
+                  outgoing=True,
+                  regex=r"red(panda|🐼)$")
+async def redpandas(event: NewMessage.Event) -> None:
+    """Get random pictures of red pandas."""
+    panda = await _request('https://some-random-api.ml/img/red_panda')
+    if not panda:
+        await event.answer("`Couldn't fetch a red panda for you :(`")
+        return
+
+    _, json = panda
+    try:
+        await event.answer(file=json['link'], reply_to=event.reply_to_msg_id)
         await event.delete()
     except rpcerrorlist.TimeoutError:
         await event.answer("`Event timed out!`")
@@ -546,7 +627,7 @@ async def waifu(event: NewMessage.Event) -> None:
                   regex="insult$")
 async def memereview(event: NewMessage.Event) -> None:
     """Insult people."""
-    await event.answer(random.choice(INSULTS))
+    await event.answer(f"__{random.choice(INSULTS)}__")
 
 
 @client.onMessage(command=("clap", plugin_category),
